+++ manager/tasks/preheat.go
@@ -2,6 +2,7 @@ package tasks
 
 import (
 	"encoding/json"
+	"errors"
 	"fmt"
 	"io/ioutil"
 	"net/http"
@@ -11,6 +12,7 @@ import (
 
 	internaltasks "d7y.io/dragonfly/v2/internal/tasks"
 	"d7y.io/dragonfly/v2/manager/types"
+	"d7y.io/dragonfly/v2/pkg/util/net/httputils"
 	machineryv1tasks "github.com/RichardKnop/machinery/v1/tasks"
 	"github.com/docker/distribution"
 	"github.com/docker/distribution/manifest/schema2"
@@ -34,7 +36,7 @@ type preheat struct {
 	url         string
 	filter      string
 	tag         string
-	headers     map[string]string
+	header      http.Header
 	tasks       *internaltasks.Tasks
 	queues      []internaltasks.Queue
 	protocol    string
@@ -42,14 +44,14 @@ type preheat struct {
 	name        string
 }
 
-func newPreheat(tasks *internaltasks.Tasks, hostnames []string, preheatType PreheatType, url, filter, tag string, headers map[string]string) Preheat {
+func newPreheat(tasks *internaltasks.Tasks, hostnames []string, preheatType PreheatType, url, filter, tag string, rawHeader map[string]string) Preheat {
 	p := &preheat{
 		tasks:       tasks,
 		preheatType: preheatType,
 		url:         url,
 		filter:      filter,
 		tag:         tag,
-		headers:     headers,
+		header:      httputils.MapToHeader(rawHeader),
 	}
 
 	for _, hostname := range hostnames {
@@ -61,19 +63,19 @@ func newPreheat(tasks *internaltasks.Tasks, hostnames []string, preheatType Preh
 		p.queues = append(p.queues, queue)
 	}
 
+	result := imageManifestsPattern.FindSubmatch([]byte(p.url))
+	if len(result) == 5 {
+		p.protocol = string(result[1])
+		p.domain = string(result[2])
+		p.name = string(result[3])
+	}
+
 	return p
 }
 
 func (p *preheat) CreatePreheat() (*types.Preheat, error) {
 	if p.preheatType == PreheatImageType {
-		result := imageManifestsPattern.FindSubmatch([]byte(p.url))
-		if len(result) == 5 {
-			p.protocol = string(result[1])
-			p.domain = string(result[2])
-			p.name = string(result[3])
-		}
-
-		layers, err := p.getLayers(p.url, p.headers, true)
+		layers, err := p.getLayers(p.url, p.header)
 		if err != nil {
 			return nil, err
 		}
@@ -86,7 +88,7 @@ func (p *preheat) CreatePreheat() (*types.Preheat, error) {
 			URL:     p.url,
 			Tag:     p.tag,
 			Filter:  p.filter,
-			Headers: p.headers,
+			Headers: httputils.HeaderToMap(p.header),
 		},
 	})
 }
@@ -124,63 +126,91 @@ func (p *preheat) createPreheat(files []*internaltasks.PreheatRequest) (*types.P
 	}, nil
 }
 
-func (p *preheat) getLayers(url string, header map[string]string, retryIfUnAuth bool) (layers []*internaltasks.PreheatRequest, err error) {
-	req, err := http.NewRequest("GET", url, nil)
+func (p *preheat) getLayers(url string, header http.Header) ([]*internaltasks.PreheatRequest, error) {
+	resp, err := p.getManifests(url, header)
 	if err != nil {
-		return
-	}
-	for k, v := range header {
-		req.Header.Add(k, v)
-	}
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return
+		return nil, err
 	}
 	defer resp.Body.Close()
 
-	body, _ := ioutil.ReadAll(resp.Body)
 	if resp.StatusCode/100 != 2 {
-		if retryIfUnAuth {
+		if resp.StatusCode == http.StatusUnauthorized {
 			token := p.getAuthToken(resp.Header)
-			if token != "" {
-				authHeader := map[string]string{
-					"Authorization": "Bearer " + token,
-					"Accept":        schema2.MediaTypeManifest,
-				}
-				return p.getLayers(url, authHeader, false)
+			bearer := "Bearer " + token
+			header.Add("Authorization", bearer)
+
+			resp, err = p.getManifests(url, header)
+			if err != nil {
+				return nil, err
 			}
+		} else {
+			return nil, errors.New(fmt.Sprintf("request registry %d", resp.StatusCode))
 		}
-		err = fmt.Errorf("%s %s", resp.Status, string(body))
-		return
 	}
 
-	layers, err = p.parseLayers(body, header)
-	return
+	layers, err := p.parseLayers(resp, header)
+	if err != nil {
+		return nil, err
+	}
+
+	return layers, nil
+}
+
+func (p *preheat) getManifests(url string, header http.Header) (*http.Response, error) {
+	req, err := http.NewRequest("GET", url, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	req.Header = header
+	req.Header.Add("Accept", schema2.MediaTypeManifest)
+
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		return nil, err
+	}
+
+	return resp, nil
+}
+
+func (p *preheat) parseLayers(resp *http.Response, header http.Header) ([]*internaltasks.PreheatRequest, error) {
+	body, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	manifest, _, err := distribution.UnmarshalManifest(schema2.MediaTypeManifest, body)
+	if err != nil {
+		return nil, err
+	}
+
+	var layers []*internaltasks.PreheatRequest
+	for _, v := range manifest.References() {
+		digest := v.Digest.String()
+		layers = append(layers, &internaltasks.PreheatRequest{
+			URL:     p.layerURL(digest),
+			Tag:     p.tag,
+			Filter:  p.filter,
+			Digest:  digest,
+			Headers: httputils.HeaderToMap(header),
+		})
+	}
+
+	return layers, nil
 }
 
 func (p *preheat) getAuthToken(header http.Header) (token string) {
-	if len(header) == 0 {
-		return
-	}
-	var values []string
-	for k, v := range header {
-		if strings.ToLower(k) == "www-authenticate" {
-			values = v
-		}
-	}
-	if values == nil {
-		return
-	}
-	authURL := p.authURL(values)
+	authURL := p.authURL(header.Values("WWW-Authenticate"))
 	if len(authURL) == 0 {
 		return
 	}
+
 	resp, err := http.Get(authURL)
 	if err != nil {
 		return
 	}
 	defer resp.Body.Close()
+
 	body, _ := ioutil.ReadAll(resp.Body)
 	var result map[string]interface{}
 	json.Unmarshal(body, &result)
@@ -203,27 +233,6 @@ func (p *preheat) authURL(wwwAuth []string) string {
 	host := strings.Split(fileds[0], "=")[1]
 	query := strings.Join(fileds[1:], "&")
 	return fmt.Sprintf("%s?%s", host, query)
-}
-
-func (p *preheat) parseLayers(body []byte, header map[string]string) ([]*internaltasks.PreheatRequest, error) {
-	manifest, _, err := distribution.UnmarshalManifest(schema2.MediaTypeManifest, body)
-	if err != nil {
-		return nil, err
-	}
-
-	var layers []*internaltasks.PreheatRequest
-	for _, v := range manifest.References() {
-		digest := v.Digest.String()
-		layers = append(layers, &internaltasks.PreheatRequest{
-			URL:     p.layerURL(digest),
-			Tag:     p.tag,
-			Filter:  p.filter,
-			Digest:  digest,
-			Headers: header,
-		})
-	}
-
-	return layers, nil
 }
 
 func (p *preheat) manifestURL(digest string) string {
